Split metadata handling out of firmware
There's already a lot in there
